# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aldeias Games API",
    "description": "API para plataforma SaaS multi-tenant de angariacao de fundos em comunidades locais portuguesas",
    "version": "3.12.0",
    "contact": {
      "name": "Sandro Pereira",
      "url": "https://github.com/smpsandro1239/Aldeias_Games"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://aldeias-games.vercel.app",
      "description": "Producao"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Obter dados do utilizador autenticado",
        "responses": {
          "200": { "description": "Dados do utilizador" },
          "401": { "description": "Nao autorizado" }
        }
      }
    },
...
  }
}